epkg update: force download by removing release file
Here the Release is outdated, leading to 404 in followup downloads and
hasher error:

wfg /c/epkg% epkg update
[00:00:03] [==========] 0 B/s        (0s) Downloaded /home/wfg/.cache/epkg/downloads/debian/dists/trixie/Release
[00:00:02] [==========] 0 B/s        (0s) Error: Err(HTTP 404 error: Client Error - https://mirrors.163.com///debian/dists/trixie/main/by-hash/SHA256/48a648f1a285637ae6153055a6430391595a9c76b4ac1abc343354e5aed9dafe

Location:
    src/download.rs:661:21)

+++ src/deb_repo.rs
@@ -117,7 +117,7 @@ pub fn parse_release_file(repo: &RepoRevise, content: &str, release_dir: &PathBu
             let parts: Vec<&str> = line.split_whitespace().collect();
             if parts.len() >= 3 {
                 let hash = parts[0].to_string();
-                let size = parts[1].parse::<u64>().unwrap_or(0);
+                let size = parts[1].parse::<usize>().unwrap_or(0);
                 let location = parts[2].to_string();
 
                 if location.contains("/debian-installer/") {
